requirements.txt
Django==3.2.12
django-admin-sortable2==1.0.4
django-tinymce==3.4.0
environs==9.5.0
Pillow==9.1.0
requests==2.27.1